backend/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
openai==1.12.0
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pytest==8.0.2
httpx==0.27.0
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
email-validator==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]>=1.7.4
python-multipart==0.0.9 
pymongo==4.11.0